fix modal issue
