+++ models/index.js
@@ -36,34 +36,4 @@ Object.keys(db).forEach(modelName => {
 db.sequelize = sequelize;
 db.Sequelize = Sequelize;
 
-/*
-//Spotify constructors
-
-var spotify = {
-  id: process.env.SPOTIFY_ID,
-  secret: process.env.SPOTIFY_SECRET
-};
-
-
-function displaySpotify() {
-  var spotify = new Spotify(config.spotify);
-  spotify.search({type: 'track', query: arg2, limit: 3}, function(error, data) {
-      if (error) {
-          return console.log("Error while retrieving a user's song: " + error);
-      } else {
-          console.log("Here are some songs that matched your search: ");
-          for (var i = 0; i <= 2; i++) {
-              var songInfo = data.tracks.items[i];
-              console.log("-----------------------------------------" +
-                          "\nArtists: " + songInfo.artists[0].name + 
-                          "\nSong: " + songInfo.name +
-                          "\nPreview: " + songInfo.external_urls.spotify +
-                          "\nAlbum: " + songInfo.album.name);
-          }
-          
-      }
-  });
-}
-*/
-
 module.exports = db;
