--- public/assets/playlist.html
@@ -1,139 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" class="meh">
-<head>
-    <meta charset="UTF-8">
-    <title>PLAYLISTED</title>
-    <!-- Added a link to Style and font awesome-->
-    <link href="../assets/css/app_style.css" rel="stylesheet" type="text/css">
-    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.2.0/css/all.css" integrity="sha384-hWVjflwFxL6sNzntih27bfxkr27PmbbK/iSvJ+a4+0owXq79v+lsFkW54bOGbiDQ" crossorigin="anonymous">
-    <!-- Font link-->
-    <link href="https://fonts.googleapis.com/css?family=Rubik" rel="stylesheet">
-    <!-- Added a link to JavaScript-->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-    <script type="text/javascript" src="../assets/js/app.js"></script>
-    <script type="text/javascript" src="../assets/js/playlist.js"></script>
-    <!-- Added a link to Bootstrap-->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-</head>
-<body>
-
-    <div class="container" id="home">
-        <div class="row">
-            <div class="col-lg-3 col-sm-12 sidebar">            
-                <div class="logo">
-                    <a href="/"><img id="logo" src="../assets/img/logo-meh.png"></a>
-                    <button class="hamburger">&#9776;</button>
-                </div>
-                <!-- SIDEBAR MENU -->
-                <div class="menu">
-                    <ul class="nav">
-                        <li>
-                            <a href="/"><i class="fa fa-home"></i>Home</a>
-                        </li>
-                        <li>
-                            <a href="#"><i class="fas fa-list-ul"></i>My Lists</a>
-                        </li>
-                        <li>
-                            <a href="/login"><i class="fas fa-sign-in-alt"></i>Login</a>
-                        </li>
-                    </ul>
-                </div>
-            </div>
-            <div class="col-lg-9 col-sm-12">
-                <div class="jumbotron">
-                    <h1 style="text-align: center; line-height: 0em; margin: 0;">PLAYLISTED <i class="fas fa-music"></i></h1>
-                    <div class="slogan">Music that fits.</div>
-                </div>
-                <div class="col-lg-12 col-sm-12 content">
-                    <div class="slidecontainer">
-                        <h3 style="text-align: center">MOOD</h3>
-                        <div class="scroll-img mood">
-                            <img id="angry" src="../assets/img/angry.png">
-                            <img id="sad" src="../assets/img/sad.png">
-                            <img id="meh" src="../assets/img/meh-color.png">
-                            <img id="happy" src="../assets/img/happy.png">
-                            <img id="ecstatic" src="../assets/img/ecstatic.png">
-                        </div>
-                        <input type="range" min="1" max="5" value="3" class="slider" id="myMood">
-                    </div>
-                    <div class="slidecontainer">
-                        <h3 style="text-align: center">ENERGY</h3>
-                        <div class="scroll-img energy">
-                            <img id="zzz" src="../assets/img/zzz-meh.png">
-                            <img id="hyper" src="../assets/img/hyper-meh.png">
-                        </div>
-                        <input type="range" min="1" max="5" value="1" class="slider" id="myEnergy">
-                    </div>
-                    <div class="genrecontainer">
-                        <h3 style="text-align: center">GENRE</h3>
-                        <form action="/action_page.php" id="formGenre">
-                            <select id="myGenre" name="genre">
-                                <option value="all" selected>All</option>
-                                <option value="classical">Classical</option>
-                                <option value="rock">Rock</option>
-                                <option value="easy listening">Easy Listening</option>
-                                <option value="pop">Pop</option>
-                                <option value="rap/hip-hop">Rap/Hip-Hop</option>
-                                <option value="unknown">Unknown</option>
-                            </select>
-                            <input type="submit" value="Listen" data-toggle="modal" data-target="#myModal" class="submitButton">
-                        </form>
-
-                        <!--container for generated playlist-->
-                        <div id="playlist-container">
-                            <!--containers for each song-->
-                            <!--This would do well as a flexbox. I will do the CSS for it with time permitting.-->
-                            <!--format: (justify left) song_title, "By: ", artist-->
-                            <!--        (justify right) favorite-button, delete-song button-->
-
-                            <div class="song-container">
-                                <!--container for favorite button-->
-                                <div class="favorite-button-container" data-commentid="1"></div>
-                                <!--container for delete-song button-->
-                                <div class="delete-button-container" data-commentid="2"></div>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="test-pull">
-                        <button id="songs1">AHHHHHH</button>
-                    </div>
-                    
-                    <!-- Modal -->
-                    <div id="myModal" class="modal" role="dialog">
-                        <div class="modal-dialog modal-md">
-                            <!-- Modal content-->
-                                <div class="modal-content">
-                                    <div class="modal-header">
-                                        <h4 class="modal-title" style="text-align:center;width:100%;">HERE'S YOUR PLAYLIST!</h4>
-                                        <button type="button" class="close" data-dismiss="modal">&times;</button>
-                                    </div>
-                                    <div class="modal-body">
-                                        <p id="mood"></p>
-                                        <p id="energy"></p>
-                                        <p id="genre"></p>
-                                    </div>
-                                    <div class="modal-body">
-                                        <p id="playlist"></p>
-                                    </div>
-                                    <div class="modal-footer">
-                                        <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <!-- Load React. -->
-    <!-- Note: when deploying, replace "development.js" with "production.min.js". -->
-    <script src="https://unpkg.com/react@16/umd/react.development.js" crossorigin></script>
-    <script src="https://unpkg.com/react-dom@16/umd/react-dom.development.js" crossorigin></script>
-
-    <!-- Load our React component. -->
-
-</body>
-</html>
