+++ views/layouts/main.handlebars
@@ -5,6 +5,7 @@
     <title>PLAYLISTED</title>
     <!-- Added a link to Style and font awesome-->
 	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
+   
 	<link rel="stylesheet" href="/assets/css/app_style.css" rel="stylesheet" type="text/css">
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.2.0/css/all.css" integrity="sha384-hWVjflwFxL6sNzntih27bfxkr27PmbbK/iSvJ+a4+0owXq79v+lsFkW54bOGbiDQ" crossorigin="anonymous">
     <!-- Font link-->
@@ -18,6 +19,7 @@
     <script type="text/javascript" src="/assets/js/register.js"></script>
     <!-- Added a link to Bootstrap-->
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+
 	<script src="https://code.jquery.com/jquery.js"></script>
 	
 </head>
@@ -66,5 +68,6 @@
             </div>
         </div>
     </div>
+
 </body>
 </html>
